Added alternative login method via Ninja server
This commit introduces an alternative way to authenticate users. If the NINJA_URL environment variable is set, the application will use a new authentication method that communicates with a Ninja server. This includes creating new files for handling Ninja server responses and errors, as well as modifying the existing login function to accommodate this change. The .env.example file has also been updated to include the NINJA_URL variable.