Added timestamp to Docker image tags
In this revision, we've introduced a change in the way we tag our Docker images. Instead of using the input image tag, we're now tagging them with the current date and time. This is achieved by setting an environment variable with the current datetime which is then used as the tag for both local and ghcr.io images. The timezone has been set to 'Asia/HongKong'.