Extended Docker build and push to 'release' branch
The Docker build and push workflow has been extended to include the 'release' branch. Now, in addition to the 'main' branch, any changes pushed to the 'release' branch will trigger this workflow. Also, a new environment variable BRANCH_NAME has been added which holds the name of the current Git branch. The login and build-push actions for ghcr are now conditional on whether BRANCH_NAME is set to 'release'. Similar changes have also been made in the Release workflow where it's now triggered by pushes to the 'release' branch instead of 'main'.